requirements.txt
alembic==1.4.2
APScheduler==3.6.3
astroid==2.4.1
Authlib==0.14.2
Bootstrap-Flask==1.3.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Deprecated==1.2.9
dominate==2.5.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
langdetect==1.0.8
lazy-object-proxy==1.4.3
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyGithub==1.51
PyJWT==1.7.1
pylint==2.5.2
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
tzlocal==2.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
